+++ src/main/kotlin/pw/vodes/rimurukt/Main.kt
@@ -6,12 +6,12 @@ import org.javacord.api.DiscordApi
 import org.javacord.api.DiscordApiBuilder
 import org.javacord.api.entity.activity.ActivityType
 import org.javacord.api.entity.server.Server
-import pw.vodes.rimurukt.audit.AuditLogs
 import pw.vodes.rimurukt.command.Commands
-import pw.vodes.rimurukt.file.AutoMod
-import pw.vodes.rimurukt.file.AutoRoles
-import pw.vodes.rimurukt.listeners.MemberListeners
-import pw.vodes.rimurukt.updater.Updater
+import pw.vodes.rimurukt.misc.MemberListeners
+import pw.vodes.rimurukt.misc.Updater
+import pw.vodes.rimurukt.services.AuditLogs
+import pw.vodes.rimurukt.services.AutoMod
+import pw.vodes.rimurukt.services.AutoRoles
 import java.io.File
 import kotlin.system.exitProcess
