+++ src/main/kotlin/pw/vodes/rimurukt/command/commands/CommandRestart.kt
@@ -5,8 +5,8 @@ import org.javacord.api.interaction.SlashCommandBuilder
 import org.javacord.api.interaction.SlashCommandInteraction
 import pw.vodes.rimurukt.command.Command
 import pw.vodes.rimurukt.command.CommandType
+import pw.vodes.rimurukt.misc.Updater
 import pw.vodes.rimurukt.reply
-import pw.vodes.rimurukt.updater.Updater
 
 class CommandRestart : Command("Restart", arrayOf("restart"), CommandType.MOD, "restart") {
     override fun run(event: MessageCreateEvent) {
